api/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
pytest==7.1.2

python-dotenv==0.20.0
pytest-cov==3.0.0

geojson==2.5.0

psycopg2-binary==2.9.3
SQLAlchemy==1.4.39
GeoAlchemy2==0.12.1

requests==2.28.1
urllib3==1.25.8

Pillow==9.2.0
shapely==1.8.2

